requirements.txt
flask==2.3.3
supabase-py==2.8.1
pandas==2.2.2
openpyxl==3.1.5
python-dotenv==1.0.1